Created Discord platform module and moved all platform-specific commands handling there out of platform class
